requirements.txt
-e git+https://github.com/patarapolw/AnkiTools.git@4424d94f775e58a9140f98b3cbd28e146bdf8a6b#egg=AnkiTools
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pyobjc==4.1
pyobjc-core==4.1
pyobjc-framework-Accounts==4.1
pyobjc-framework-AddressBook==4.1
pyobjc-framework-AppleScriptKit==4.1
pyobjc-framework-AppleScriptObjC==4.1
pyobjc-framework-ApplicationServices==4.1
pyobjc-framework-Automator==4.1
pyobjc-framework-AVFoundation==4.1
pyobjc-framework-AVKit==4.1
pyobjc-framework-CalendarStore==4.1
pyobjc-framework-CFNetwork==4.1
pyobjc-framework-CloudKit==4.1
pyobjc-framework-Cocoa==4.1
pyobjc-framework-Collaboration==4.1
pyobjc-framework-Contacts==4.1
pyobjc-framework-ContactsUI==4.1
pyobjc-framework-CoreBluetooth==4.1
pyobjc-framework-CoreData==4.1
pyobjc-framework-CoreLocation==4.1
pyobjc-framework-CoreServices==4.1
pyobjc-framework-CoreText==4.1
pyobjc-framework-CoreWLAN==4.1
pyobjc-framework-CryptoTokenKit==4.1
pyobjc-framework-DictionaryServices==4.1
pyobjc-framework-DiskArbitration==4.1
pyobjc-framework-EventKit==4.1
pyobjc-framework-ExceptionHandling==4.1
pyobjc-framework-FinderSync==4.1
pyobjc-framework-FSEvents==4.1
pyobjc-framework-GameCenter==4.1
pyobjc-framework-GameController==4.1
pyobjc-framework-GameKit==4.1
pyobjc-framework-GameplayKit==4.1
pyobjc-framework-ImageCaptureCore==4.1
pyobjc-framework-IMServicePlugIn==4.1
pyobjc-framework-InputMethodKit==4.1
pyobjc-framework-InstallerPlugins==4.1
pyobjc-framework-InstantMessage==4.1
pyobjc-framework-Intents==4.1
pyobjc-framework-IOSurface==4.1
pyobjc-framework-iTunesLibrary==4.1
pyobjc-framework-LatentSemanticMapping==4.1
pyobjc-framework-LaunchServices==4.1
pyobjc-framework-libdispatch==4.1
pyobjc-framework-LocalAuthentication==4.1
pyobjc-framework-MapKit==4.1
pyobjc-framework-MediaAccessibility==4.1
pyobjc-framework-MediaLibrary==4.1
pyobjc-framework-MediaPlayer==4.1
pyobjc-framework-ModelIO==4.1
pyobjc-framework-MultipeerConnectivity==4.1
pyobjc-framework-NetFS==4.1
pyobjc-framework-NetworkExtension==4.1
pyobjc-framework-NotificationCenter==4.1
pyobjc-framework-OpenDirectory==4.1
pyobjc-framework-Photos==4.1
pyobjc-framework-PhotosUI==4.1
pyobjc-framework-PreferencePanes==4.1
pyobjc-framework-PubSub==4.1
pyobjc-framework-QTKit==4.1
pyobjc-framework-Quartz==4.1
pyobjc-framework-SafariServices==4.1
pyobjc-framework-SceneKit==4.1
pyobjc-framework-ScreenSaver==4.1
pyobjc-framework-ScriptingBridge==4.1
pyobjc-framework-SearchKit==4.1
pyobjc-framework-Security==4.1
pyobjc-framework-SecurityFoundation==4.1
pyobjc-framework-SecurityInterface==4.1
pyobjc-framework-ServiceManagement==4.1
pyobjc-framework-Social==4.1
pyobjc-framework-SpriteKit==4.1
pyobjc-framework-StoreKit==4.1
pyobjc-framework-SyncServices==4.1
pyobjc-framework-SystemConfiguration==4.1
pyobjc-framework-WebKit==4.1
pywebview==1.8
requests==2.18.4
sip==4.19.8
urllib3==1.22
web.py==0.40.dev1
